Minimal correction

+++ src/main/java/com/distrimind/ood/OOD.java
@@ -65,7 +65,7 @@ public class OOD {
 			VERSION.addCreator(new Person("mahdjoub", "jason"))
 					.addDeveloper(new PersonDeveloper("mahdjoub", "jason", "2013-04-01"))
 					.addDescription(
-							new Description((short)3, (short)1, (short)22, Version.Type.STABLE, (short)0, "2022-04-05")
+							new Description(3, 1, 22, Version.Type.STABLE, 0, "2022-04-05")
 									.addItem(INTERNAL_CHANGE, "Update Utils to 5.23.2 STABLE")
 									.addItem(INTERNAL_CHANGE, "Add lacking files into generated libraries.")
 					)
